requirements.txt
mindspore-gpu>=1.5.0
numpy